// src/lib/api/types/reportWithContentDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */
import type { ContentDTO } from './contentDTO';
import type { ReportDTO } from './reportDTO';

export interface ReportWithContentDTO {
  /** Reported content details */
  content?: ContentDTO;
  /** Report details */
  report: ReportDTO;
}